Update auth user command & its applications

+++ main.py
@@ -58,44 +58,57 @@ task_completed = discord.Embed(
 #status button embed gui
 class Buttons(discord.ui.View):
     def __init__(self):
-        super().__init__()
+        super().__init__(timeout=0)
+
+    async def on_timeout(self):
+        pass
 
     @discord.ui.button(label="Startup", style=discord.ButtonStyle.green)
     async def start(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message(embed=task_completed, ephemeral=True)
-        channel = bot.get_channel(1222911491821277276)
-        status = bot.get_channel(1223028334170996826)
-        embed = discord.Embed(
-            title="Server Status: Online",
-            description="Server Startup! \n Join in erlc with code: NTC",
-            colour=discord.Color.green()
-        )
-        async for message in channel.history(limit=None):
-            await message.delete()
-        await status.edit(name="Server Status: Online")
-        await channel.send(embed=embed)
-        await channel.send("@everyone Server Has Started!", delete_after=5)
-        await channel.last_message.delete()
+
+        authenticated_users = await read_authenticated_users()
+
+        if interaction.user.id in authenticated_users:
+            await interaction.response.send_message(embed=task_completed, ephemeral=True)
+            channel = bot.get_channel(1222911491821277276)
+            status = bot.get_channel(1223028334170996826)
+            embed = discord.Embed(
+                title="Server Status: Online",
+                description="Server Startup! \n Join in erlc with code: NTC",
+                colour=discord.Color.green()
+            )
+            async for message in channel.history(limit=None):
+                await message.delete()
+            await status.edit(name="Server Status: Online")
+            await channel.send(embed=embed)
+            await channel.send("@everyone Server Has Started!", delete_after=5)
+            await channel.last_message.delete()
+        else:
+            await interaction.response.send_message(embed=perm_error, ephemeral=True)
+        
 
     @discord.ui.button(label="Shutdown", style=discord.ButtonStyle.danger)
     async def shutdown(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message(embed=task_completed, ephemeral=True)
-        channel = bot.get_channel(1222911491821277276)
-        status = bot.get_channel(1223028334170996826)
-        embed = discord.Embed(
-            title="Server Status: Offline",
-            description="Server Is Offline.",
-            colour=discord.Color.red()
-        )
-        async for message in channel.history(limit=None):
-            await message.delete()
-        await status.edit(name="Server Status: Offline")
-        await channel.send(embed=embed)
 
+        authenticated_users = await read_authenticated_users()
 
-
-
-
+        if interaction.user.id in authenticated_users:
+            await interaction.response.send_message(embed=task_completed, ephemeral=True)
+            channel = bot.get_channel(1222911491821277276)
+            status = bot.get_channel(1223028334170996826)
+            embed = discord.Embed(
+                title="Server Status: Offline",
+                description="Server Is Offline.",
+                colour=discord.Color.red()
+            )
+            async for message in channel.history(limit=None):
+                await message.delete()
+            await status.edit(name="Server Status: Offline")
+            await channel.send(embed=embed)
+        else:
+            await interaction.response.send_message(embed=perm_error, ephemeral=True)
+        
+        
 # apply for staff button embed gui
 class Jobs(discord.ui.View):
     def __init__(self):
@@ -340,6 +353,7 @@ async def on_ready():
     print(f'Logged in as {bot.user}')
     print(bot.guilds)
     await support_reset_startup()
+    await SSU_change_reset_startup()
     #subprocess.Popen(["python", "backend.py"])
     try:
         synced = await bot.tree.sync()
@@ -433,6 +447,21 @@ async def support_reset_startup():
 
         await support_channel.send(embed=embed, view=view)
 
+
+# sets the support message
+async def SSU_change_reset_startup():
+    guild = bot.guilds[0]
+    support_channel = guild.get_channel(1223697564390522951)
+    view = Buttons()
+
+    await support_channel.purge(limit=None)
+
+    embed = discord.Embed(
+        title="Server Status",
+        description="Do you want to start or stop the server?",
+        color=discord.Color.blue()
+    )
+    await support_channel.send(embed=embed, view=view)
 
 
 
@@ -590,9 +619,11 @@ async def announcement(interaction: discord.Interaction):
 async def status(interaction: discord.Interaction):
     owner_role = discord.utils.get(interaction.guild.roles, name="Owner")
     authenticated_users = await read_authenticated_users()
+    guild = bot.guilds[0]
+    support_channel = guild.get_channel(1223697564390522951)
 
     if interaction.user.id in authenticated_users and owner_role in interaction.user.roles:
-        await interaction.response.send_message(embed=sent_dm, ephemeral=True)
+        await support_channel.purge(limit=None)
 
         embed = discord.Embed(
             title="Server Status",
@@ -600,42 +631,39 @@ async def status(interaction: discord.Interaction):
             color=discord.Color.blue()
         )
         buttons = Buttons()
-        await interaction.user.send(embed=embed, view=buttons)
+        await support_channel.send(embed=embed, view=buttons)
 
 
 
 
 # auth slash command
 @bot.tree.command(name="auth", description="add users to auth list")
-async def auth(interaction: discord.Interaction):
+async def auth(interaction: discord.Interaction, user: discord.Member):
     supervisor_role = discord.utils.get(interaction.guild.roles, name="Owner")
     authenticated_users = await read_authenticated_users()
 
-    if supervisor_role in interaction.user.roles and int(interaction.user.id) in authenticated_users:
+    if supervisor_role in user.roles and int(user.id) in authenticated_users:
         embed = discord.Embed(
             title="New Town City Support",
-            description="You are already authenticated!",
+            description=f"{user.name} is already authenticated!",
             color=discord.Color.yellow()  # You can customize the color if needed
         )
-        await interaction.user.send(embed=embed)
-        await interaction.response.send_message(embed=sent_dm, ephemeral=True)
+        await interaction.response.send_message(embed=embed, ephemeral=True)
     elif supervisor_role in interaction.user.roles:
         # User has the "supervisor" role, proceed with the command
         embed = discord.Embed(
             title="New Town City Support",
-            description="You have authenticated with NTC support system!",
+            description=f"{user.name} has been authenticated with NTC support system!",
             color=discord.Color.green()  # You can customize the color if needed
         )
-        await interaction.user.send(embed=embed)
-        await interaction.response.send_message(embed=sent_dm, ephemeral=True)
+        await interaction.response.send_message(embed=embed, ephemeral=True)
 
-        authenticated_users.append(interaction.user.id)
-        print(f"Debug: userid is ", interaction.user.id)
+        authenticated_users.append(user.id)
+        print(f"Debug: userid is ", user.id)
         await write_authenticated_users(authenticated_users)
     else:
         # User does not have the required role, send a message indicating the restriction
         await interaction.response.send_message(embed=perm_error, ephemeral=True)
-        await interaction.user.send(embed=perm_error)
 
 
 
@@ -643,15 +671,14 @@ async def auth(interaction: discord.Interaction):
 
 # remove slash command
 @bot.tree.command(name="remove", description="Remove user IDs from auth")
-async def remove(interaction: discord.Interaction, userid: str):
+async def remove(interaction: discord.Interaction, user: discord.Member):
     owner_role = 618331718692241408
-    print("userid is ", userid)
+    print("userid is ", user.id)
 
     try:
         authenticated_users = await read_authenticated_users()
-        if owner_role in interaction.user.id and int(userid) in authenticated_users:
-            await interaction.response.send_message(embed=sent_dm, ephemeral=True)
-            authenticated_users.remove(int(userid))
+        if owner_role in interaction.user.id and int(user.id) in authenticated_users:
+            authenticated_users.remove(int(user.id))
             await write_authenticated_users(authenticated_users)
 
             confirmation_embed = discord.Embed(
@@ -659,19 +686,17 @@ async def remove(interaction: discord.Interaction, userid: str):
                 description=f"User with ID {userid} was removed from authenticated users.",
                 color=discord.Color.green()
             )
-            await interaction.user.send(embed=confirmation_embed)
+            await interaction.response.send_message(embed=confirmation_embed)
 
-        elif owner_role in interaction.user.roles:
+        elif owner_role in interaction.user.id:
             not_found_embed = discord.Embed(
                 title="Error",
-                description=f"User with ID {userid} is not in the list of authenticated users.",
+                description=f"User with ID {user.id} is not in the list of authenticated users.",
                 colour=discord.Color.red()
             )
-            await interaction.response.send_message(embed=sent_dm, ephemeral=True)
-            await interaction.user.send(embed=not_found_embed)
+            await interaction.response.send_message(embed=not_found_embed, ephemeral=True)
         else:
             await interaction.response.send_message(embed=perm_error, ephemeral=True)
-            await interaction.user.send(embed=perm_error)
 
     except Exception as e:
         print(e)
@@ -687,9 +712,7 @@ async def help(interaction: discord.Interaction):
         description="/help (shows the list of commands) \n /auth (authenticates users to use NTC commands) \n /remove (removes users from the auth list) \n /announcement (posts an announcement in the channel) \n /support_rest (resets the ticket support sys) \n /reload_cfg (reloads the bots config) \n /status (changes the servers status) \n /close (closes the ticket) \n /claim (claims the ticket)",
         color=discord.Color.blue()
     )
-    user=interaction.user
-    await user.send(embed=commands_embed)
-    await interaction.response.send_message(embed=sent_dm, ephemeral=True)
+    await interaction.response.send_message(embed=commands_embed, ephemeral=True)
 
 
 # reload_cfg slash command
